Mostly fix issue #2
Rackproxy can now load plugins on Windows.

The issue concerning missing exported symbols was solved by staring at
the Rack Makefile :-)

The only thing left is, that since Rack plugins on Windows have a
hard-dependency on "rack.exe", the executable needs to be named rack.exe
rather than rackproxy.exe on Windows. It would be nice to "fix" this, if
at all possible.